requirements.txt
python-dotenv==0.10.3
requests==2.22.0
vk-api==11.5.1
python-telegram-bot==12.2.0
pydrive==1.3.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1